Mark as Read button fully functional
Many things had to be changed, so check for merge conflicts. Global varibles needed to be set for readButton, deleteButton, websiteTab. CSS was changed and added. .website-tab-buttons changed background color to inherit. class read width changes. read-button-read was created. that name is aweful. Too tired to think of better name.